+++ src/hwenc_nvcodec/nvcodec_video_encoder.cpp
@@ -20,6 +20,8 @@
 #include <modules/video_coding/codecs/h264/include/h264.h>
 #include <modules/video_coding/include/video_codec_interface.h>
 #include <modules/video_coding/include/video_error_codes.h>
+#include <modules/video_coding/svc/create_scalability_structure.h>
+#include <modules/video_coding/svc/scalable_video_controller.h>
 #include <rtc_base/logging.h>
 
 // libyuv
@@ -125,6 +127,10 @@ class NvCodecVideoEncoderImpl : public NvCodecVideoEncoder {
   NV_ENC_INITIALIZE_PARAMS initialize_params_;
   std::vector<std::vector<uint8_t>> v_packet_;
   webrtc::EncodedImage encoded_image_;
+
+  // AV1 用
+  std::unique_ptr<webrtc::ScalableVideoController> svc_controller_;
+  webrtc::ScalabilityMode scalability_mode_;
 };
 
 NvCodecVideoEncoderImpl::NvCodecVideoEncoderImpl(
@@ -178,6 +184,18 @@ int32_t NvCodecVideoEncoderImpl::InitEncode(
   mode_ = codec_settings->mode;
 
   RTC_LOG(LS_INFO) << "InitEncode " << target_bitrate_bps_ << "bit/sec";
+
+  if (codec_settings->codecType == webrtc::kVideoCodecAV1) {
+    auto scalability_mode = codec_settings->GetScalabilityMode();
+    if (!scalability_mode) {
+      RTC_LOG(LS_WARNING) << "Scalability mode is not set, using 'L1T1'.";
+      scalability_mode = webrtc::ScalabilityMode::kL1T1;
+    }
+    RTC_LOG(LS_INFO) << "InitEncode scalability_mode:"
+                     << (int)*scalability_mode;
+    svc_controller_ = webrtc::CreateScalabilityStructure(*scalability_mode);
+    scalability_mode_ = *scalability_mode;
+  }
 
   return InitNvEnc();
 }
@@ -341,8 +359,19 @@ int32_t NvCodecVideoEncoderImpl::Encode(
   }
 
   for (std::vector<uint8_t>& packet : v_packet_) {
-    auto encoded_image_buffer =
-        webrtc::EncodedImageBuffer::Create(packet.data(), packet.size());
+    uint8_t* p = packet.data();
+    size_t size = packet.size();
+    if (codec_ == CudaVideoCodec::AV1) {
+      // IVF ヘッダーが付いてるので取り除く
+      if ((p[0] == 'D') && (p[1] == 'K') && (p[2] == 'I') && (p[3] == 'F')) {
+        p += 32;
+        size -= 32;
+      }
+      p += 12;
+      size -= 12;
+    }
+    auto encoded_image_buffer = webrtc::EncodedImageBuffer::Create(p, size);
+
     encoded_image_.SetEncodedData(encoded_image_buffer);
     encoded_image_._encodedWidth = width_;
     encoded_image_._encodedHeight = height_;
@@ -358,22 +387,30 @@ int32_t NvCodecVideoEncoderImpl::Encode(
     encoded_image_.SetColorSpace(frame.color_space());
     encoded_image_._frameType = webrtc::VideoFrameType::kVideoFrameDelta;
 
-    uint8_t zero_count = 0;
-    size_t nal_start_idx = 0;
-    for (size_t i = 0; i < packet.size(); i++) {
-      uint8_t data = packet.data()[i];
-      if ((i != 0) && (i == nal_start_idx)) {
-        if ((data & 0x1F) == 0x05) {
-          encoded_image_._frameType = webrtc::VideoFrameType::kVideoFrameKey;
+    if (codec_ == CudaVideoCodec::H264 || codec_ == CudaVideoCodec::H265) {
+      uint8_t zero_count = 0;
+      size_t nal_start_idx = 0;
+      for (size_t i = 0; i < packet.size(); i++) {
+        uint8_t data = packet.data()[i];
+        if ((i != 0) && (i == nal_start_idx)) {
+          if ((data & 0x1F) == 0x05) {
+            encoded_image_._frameType = webrtc::VideoFrameType::kVideoFrameKey;
+          }
+        }
+        if (data == 0x01 && zero_count >= 2) {
+          nal_start_idx = i + 1;
+        }
+        if (data == 0x00) {
+          zero_count++;
+        } else {
+          zero_count = 0;
         }
       }
-      if (data == 0x01 && zero_count >= 2) {
-        nal_start_idx = i + 1;
-      }
-      if (data == 0x00) {
-        zero_count++;
-      } else {
-        zero_count = 0;
+    } else if (codec_ == CudaVideoCodec::AV1) {
+      // 最初の 2 バイトは 0x12 0x00 で、これは OBU_TEMPORAL_DELIMITER なのでこれは無視して、その次の OBU を見る
+      // 0x0a は OBU_SEQUENCE_HEADER なのでキーフレームと判断する
+      if (p[2] == 0x0a) {
+        encoded_image_._frameType = webrtc::VideoFrameType::kVideoFrameKey;
       }
     }
 
@@ -390,6 +427,24 @@ int32_t NvCodecVideoEncoderImpl::Encode(
 
       h265_bitstream_parser_.ParseBitstream(encoded_image_);
       encoded_image_.qp_ = h265_bitstream_parser_.GetLastSliceQp().value_or(-1);
+    } else if (codec_ == CudaVideoCodec::AV1) {
+      codec_specific.codecType = webrtc::kVideoCodecAV1;
+
+      bool is_key =
+          encoded_image_._frameType == webrtc::VideoFrameType::kVideoFrameKey;
+      std::vector<webrtc::ScalableVideoController::LayerFrameConfig>
+          layer_frames = svc_controller_->NextFrameConfig(is_key);
+      codec_specific.end_of_picture = true;
+      codec_specific.scalability_mode = scalability_mode_;
+      codec_specific.generic_frame_info =
+          svc_controller_->OnEncodeDone(layer_frames[0]);
+      if (is_key && codec_specific.generic_frame_info) {
+        codec_specific.template_structure =
+            svc_controller_->DependencyStructure();
+        auto& resolutions = codec_specific.template_structure->resolutions;
+        resolutions = {webrtc::RenderResolution(encoded_image_._encodedWidth,
+                                                encoded_image_._encodedHeight)};
+      }
     }
 
     webrtc::EncodedImageCallback::Result result =
@@ -415,6 +470,10 @@ void NvCodecVideoEncoderImpl::SetRates(
   if (parameters.framerate_fps < 1.0) {
     RTC_LOG(LS_WARNING) << "Invalid frame rate: " << parameters.framerate_fps;
     return;
+  }
+
+  if (svc_controller_) {
+    svc_controller_->OnRatesUpdated(parameters.bitrate);
   }
 
   uint32_t new_framerate = (uint32_t)parameters.framerate_fps;
@@ -544,6 +603,10 @@ std::unique_ptr<NvEncoder> NvCodecVideoEncoderImpl::CreateEncoder(
       encoder->CreateDefaultEncoderParams(
           &initialize_params, NV_ENC_CODEC_HEVC_GUID, NV_ENC_PRESET_P2_GUID,
           NV_ENC_TUNING_INFO_LOW_LATENCY);
+    } else if (codec == CudaVideoCodec::AV1) {
+      encoder->CreateDefaultEncoderParams(
+          &initialize_params, NV_ENC_CODEC_AV1_GUID, NV_ENC_PRESET_P2_GUID,
+          NV_ENC_TUNING_INFO_LOW_LATENCY);
     }
 
     //initialize_params.enablePTD = 1;
@@ -582,6 +645,10 @@ std::unique_ptr<NvEncoder> NvCodecVideoEncoderImpl::CreateEncoder(
       encode_config.encodeCodecConfig.hevcConfig.repeatSPSPPS = 1;
       encode_config.encodeCodecConfig.hevcConfig.sliceMode = 0;
       encode_config.encodeCodecConfig.hevcConfig.sliceModeData = 0;
+    } else if (codec == CudaVideoCodec::AV1) {
+      encode_config.encodeCodecConfig.av1Config.idrPeriod =
+          NVENC_INFINITE_GOPLENGTH;
+      encode_config.encodeCodecConfig.av1Config.repeatSeqHdr = 1;
     }
 
     encoder->CreateEncoder(&initialize_params);